requirements.txt
langchain
langchain-community
python-dotenv
requests
